requirements.txt
argparse
datetime
json
os
numpy
pandas
matplotlib
scikit-image

torch
torchvision
torchsummary
tensorboardX

Pillow 
